requirements.txt
Pillow==7.1.2
numpy==1.21.6
scikit-image==0.18.3
opencv-python==4.6.0.66
tensorflow==2.9.2